Refactor ant code and add visuals
